Remove unused/supported upstream modes
We now only support exploding mode.  Operation on groups is still
possible via `nodeattr`.  E.g., the client can be called as `flight
power status -g $(nodeattr -q compute)`, to get the power status for all
compute nodes.

If an enhancement is required to support that same functionality with
the syntax `flight power status -g compute`, we can either add suppot
for calling `nodeattr` to the client or look into adding support for
calling `nodeattr` to the server.