Remove scary and false warning in application.yaml
It's false in the sense that the functionality it warned about was never
implemented.

I'm not convinced that Figaro is a good choice for managing our
configurations.  It has the following behaviour:

 1. Configuration is set in a YAML file. But the values are restricted
    to scalar values.  And they are all converted to strings.  But
    nested values are supported for setting values for different
    environments.  YAML is a poor choice for this set of restrictions.

 2. The application.yaml file is not supposed to be committed to the
    repo as it may contain sensitive values.  It potentially contains
    complex configuration for different environments much of which is
    not sensitive.  This results in the case we have with this project.
    A configuration file that has been committed to the repo with a
    warning that sensitive values should not be added.

 3. Only a single configuration file is supported (which is not supposed
    to be committed to the repo).  Following Figaro's guidance would
    require the deployment process to recreate the complex configuration
    file.  As none of this is supposed to be committed to the repo, the
    deployment process needs to know the defaults.  Or we don't follow
    Figaro's process and commit `application.yaml` to the repo as we've
    done here.

I would imagine that Figaro works well for deployments to heroku where
the deployment typically takes place from a developers machine and the
developer can use Figaro to set the configuration in heroku.

However, for our use Figaro seems like a poor match.  I think we'd be
better off switching to something like `dotenv`.  Then we can have our
non-sensitive data set in, say, `.env` and use `.env.local` for
sensitive information.  A `.gitignore` rule would make sure that
`.env.local` is never committed.

This commit doesn't go that far.  Instead it removes the scary warning
so that `flight service configure` can be used to set the JWT secret.