Fix output sometimes being dropped
Previously, `Subprocess#run` assumed that once the process had completed
we had finished dealing with its output and it was safe to return.  On
its return, `Job#run` would close its `@write_pipe`.

However, the processing of the jobs output is performed in background
threads and there are mutexes involved. This means the assumption that
the process having completed implies its output has been processed is
wrong.  The result being that `Job#run` closed its `@write_pipe` too
early; the block given to `Subprocess#run` may still attempt to write to
it.

When that happened the, output from the job was lost.

The solution is to ensure that `Subprocess#run` doesn't return until 1)
the process has completed; and 2) its output handling threads have
finished.