Refactor validate_machine_type to support filtering
Not all the types will be implemented on every platform. The
non-implemented types need to be filter out from the error mesages

Also azure does not support arbitrarily resizing a node to a different
type. Instead an optional filter has been added to only allow resizing
within a single family